Create AptComplexV1
My second JSON development and my first on GitHub. Learning two things at once is always interesting. I am sure at one point I will look back and have a chuckle. 
In any case, trying to figure out what is the best way to add the apartment complex variations ( one, two, three bedroom etc) - as an extension of Place  as a type of offer...or as an extension to Thing, as an additional type. Will read about it.